Exported simulations configurations to separate 'restrictions' module, renaming module: 'parameter' -> 'simulation'
