+++ alcor/models/__init__.py
@@ -1,3 +1,2 @@
-from .parameter import Parameter
 from .star import (STAR_PARAMETERS_NAMES,
                    Star)
