+++ alcor/services/restrictions.py
@@ -1,0 +1,2 @@
+OUTPUT_FILE_EXTENSION = '.res'
+MAX_OUTPUT_FILE_NAME_LENGTH = 5
